Feat order by distance

+++ src/main/java/org/omsf/store/model/StoreDistance.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class StoreDistance extends Store {
 	
-	private Double distance;
+	private Integer distance;
 
 }
